:tractor: Move files around for a slightly cleaner package structure
