+++ pkg/server/web/src/templates/foo.tmpl.md
